observability/open-telemetry-example/app2/requirements.txt
# Flask==2.1.1
# opentelemetry-distro
# requests==2.27.1

Flask==2.1.1
flask-cors==3.0.10
opentelemetry-api==1.8.0
opentelemetry-sdk==1.8.0
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
requests==2.27.1
Werkzeug==2.2.2
protobuf==3.20.3